xlsbuilder: Fix row length error if excels contains more than a short (32k) rows count

+++ libraries/xlsbuilder/src/io/intino/alexandria/office/XlsBuilder.java
@@ -102,7 +102,7 @@ public class XlsBuilder {
 			if (line.startsWith("#define")) {
 				define(sheet, line);
 			} else {
-				create(sheet.createRow((short) i), line.split(";"));
+				create(sheet.createRow(i), line.split(";"));
 			}
 		}
 	}
